move serve filter utilities into a separate package
