+++ filters/serve/serve_test.go
@@ -1,4 +1,4 @@
-package builtin
+package serve
 
 import (
 	"errors"
